DB Connection, DAO's, and Class files
The DB Connection has been tested, and is working with the appropriate Maven dependency.  The DAO's have an interface they are based on, but they are not built out and tested.  The Class files have been created, but they have not been tested.  The relevent SQL database has not been created; this should be done to verify the connection between data in the database and the Java program.