Updated Long/Int and Service Layer
There was a problem where the database was using bigints for IDs and the entities were using ints. I had to convert and make sure the converts were safe, so i updated everything in java to be longs. The service layers have a structure in place, but no testing yet.